requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.11
jinja2>=3.1.0
premailer>=3.10.0
click>=8.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
sendgrid>=6.11.0
schedule>=1.2.0
